+++ src/Components/Partials/Navigation.php
@@ -6,9 +6,13 @@ use Illuminate\View\Component;
 
 class Navigation extends Component
 {
-    public function __construct()
+    public $mainMenu, $subMenu;
+
+    public function __construct($mainMenu = '', $subMenu = '')
     {
-        //
+        $this->mainMenu = $mainMenu;
+
+        $this->subMenu = $subMenu;
     }
 
     public function render()